Model Filter Bug Fix
When a user turned off a model, the update data function would fail
because that series no longer existed. The solution was simply to check
whether or not it was defined before updating it.

if (typeof chart.series[chartSeriesIndex].data[j] != 'undefined') {
chart.series[chartSeriesIndex].data[j].update(dataToPlot, false);
}