Climate Data links to Data Basin
Made modifications to the way the legend builds a link to the
corresponding dataset on Data Basin. Model names in the swapLegend()
function in leaflet_map.js now come from the config file rather than the
search and replace hack. This model name is also used to find the
DataBasin Layer Index (a new parameter in the config file).

point_chart.js now passes the model name to the swapLegend() function.